BioRegElePrioritization {NetLCP}	R Documentation

Regulatory Biological Elements Prioritization

Description

Prioritize the biological regulatory elements through decoding the heterogeneous network according to input transcriptome.

Usage

BioRegElePrioritization(transcriptomeList, prioType, empiricalPvalue)

Arguments

transcriptomeList
a vector consist transcriptome including circRNA, lncRNA, miRNA or mRNA.

prioType
character string naming the prioritizing element types including circRNA, lncRNA, KEGG, Reactome, Wikipathway.

empiricalPvalue
logical value enabling or not enabling empirical p value computing.

Value

a dataframe of prioritizations.

Examples

BioRegElePrioritization(transcriptomeList = c("2309", "3838", "MIMAT0000255"),
                                  prioType = "KEGG",
                                  empiricalPvalue = FALSE)

[Package NetLCP version 1.0.0 Index]