binaryRegulation {NetLCP}	R Documentation

Binary Regulatory in Local Heterogeneous network

Description

Locate the experimental binary regulatory data among biological elements in regional Heterogeneous network which is integrated from a wide range of bioinformatics databases.

Usage

binaryRegulation(transcriptomeList, regulationType, allRegulation)

Arguments

transcriptomeList
a vector consist of transcriptome including circRNA, lncRNA, miRNA ,mRNA or KEGG/Reactome/Wikipathway pathway IDs.

regulationType
character string naming the type of regulation which can be located including "circRNA-miRNA", "lncRNA-miRNA", "lncRNA-mRNA", "miRNA-mRNA", "miRNA-pathway", "mRNA-pathway".

allRegulation
a logical value indicates locating all related regulations from the storage or internal regulations between input elements.

Value

The dataframe of regulatory data with the colnames of Node1, Node2, Datasoure and RegulationType.

Examples

binaryRegulation(transcriptomeList = c("2309", "3838", "MIMAT0000255"), regulationType = "miRNA-mRNA",  allRegulation = FALSE)

[Package NetLCP version 1.0.0 Index]