eQTLsRegStat {NetLCP}	R Documentation

eQTLs on regulations statistics

Description

eQTLs on regulations statistics

Usage

eQTLsRegStat(regData, eQTLsData, regulationType, filterDegree, , selectNode)

Arguments

regData
the standard output of the function binaryRegulationExtract or multieleRegulation.

eQTLsData
the standard out of the function eQTLsDetection.

regulationType
a character representing the extracted regulation data type, it should be one of "circRNA-miRNA", "lncRNA-miRNA", "lncRNA-mRNA", "miRNA-mRNA", "miRNA-pathway", "mRNA-pathway", "circRNA-miRNA-mRNA", "lncRNA-miRNA-mRNA", "miRNA-mRNA-pathway", "lncRNA-miRNA-mRNA-pathway", "circRNA-miRNA-mRNA-pathway".

filterDegree
an integer for filtering the nodes.

selectNode
select a or a set of nodes to perform statistics.

Value

eQTLs on regulations statistics with barplot.

Examples

 eQTLsRegStat(regData = NULL, eQTLsData = NULL, regulationType = "miRNA-mRNA", filterDegree = 40, selectNode = NULL)

[Package NetLCP version 1.0.0 Index]