multieleRegulation {NetLCP}	R Documentation

Multielements Regulatory Data Extracting

Description

Locate the experimental Multielement regulatory data among biological elements in regional Heterogeneous network which is integrated from a wide range of bioinformatics databases.

Usage

multieleRegulation(transcriptomeList, regulationType, allRegulation)

Arguments

transcriptomeList
a vector consist of transcriptome including circRNA, lncRNA, miRNA, mRNA or pathway IDs.

regulationType
character string naming the type of regulation which can be located including "circRNA-miRNA-mRNA", "lncRNA-miRNA-mRNA", "lncRNA-miRNA-pathway" and "circRNA-miRNA-pathway".

allRegulation
a logical value indicates locating all related regulations from the storage or internal regulations between input elements.

Value

The dataframe of regulatory data with the colnames of Node1, Node2, Datasoure and RegulationType.

Examples

multieleRegulation(transcriptomeList = c("2309", "3838", "ENSG00000259366", "MIMAT0000255"), regulationType = "lncRNA-miRNA-mRNA",  allRegulation = FALSE)

[Package NetLCP version 1.0.0 Index]