regNetVis {NetLCP}	R Documentation

regulatory Network Visual Function

Description

Regulatory network visualization based on the standard output of the function binaryRegulation or multieleRegulation.

Usage

regNetVis(regData, filterDegree, selectNode, netLayout)

Arguments

regData
regData standard output of the function binaryRegulation or multieleRegulation.

filterDegree
an integer for filtering the nodes.

selectNode
select a or a set of nodes to visulize in the network.

netLayout
the layout of network.

Value

regulatory network visualization

Examples

 regNetVis(regData = NULL, filterDegree = 20, selectNode = "9978", netLayout = "layout_in_circle")

[Package NetLCP version 1.0.0 Index]