regVarDetection {NetLCP}	R Documentation

Mutations on regulations Detection

Description

Detect the Mutations on regulations, which are integrated from LnCeVar and miRNASNPv3

Usage

regVarDetection(regVar, regulationType)

Arguments

regData
the dataframe of output from binaryRegulation and multieleRegulation functions, accepting miRNA-mRNA, miRNA-mRNA-pathway, lncRNA-miRNA-mRNA, circRNA-miRNA-mRNA, lncRNA-miRNA-mRNA-pathway, circRNA-miRNA-mRNA-pathway.

regulationType
a character representing the extracted regulation data type, it should be "miRNA-mRNA", "miRNA-mRNA-pathway", "lncRNA-miRNA-mRNA", "circRNA-miRNA-mRNA", "lncRNA-miRNA-mRNA-pathway" or "circRNA-miRNA-mRNA-pathway".

Value

Regulatory Variants Detection.

Examples

regVarDetection(regVar = NULL, regulationType = "lncRNA-miRNA-mRNA", selectNode = sampleData)

[Package NetLCP version 1.0.0 Index]