regVarStat {NetLCP}	R Documentation

Regulatory mutation Statistics

Description

Regulatory mutation Statistics

Usage

regVarStat(regVar, regulationType, selectNode)

Arguments

regVar
the standard out of the function regVarDetection.

regulationType
a character representing the regulation data type, it should be "miRNA-mRNA", "miRNA-mRNA-pathway", "lncRNA-miRNA-mRNA", "circRNA-miRNA-mRNA", "lncRNA-miRNA-mRNA-pathway" or "circRNA-miRNA-mRNA-pathway".

selectNode
select a or a set of nodes to visulize in the network.

Value

Regulatory mutation Statistics

Examples

regVarStat(regVar = NULL, regulationType = "lncRNA-miRNA-mRNA", selectNode = sampleData)

[Package NetLCP version 1.0.0 Index]